+++ database/src/database/mod.rs
@@ -1,6 +1,7 @@
 pub mod commands;
 pub mod control;
 pub mod database;
+pub mod options;
 pub mod orchestrator;
 pub mod request_manager;
 pub mod table;
